requirements.txt
# ============================================================================
# MediBot Analysis - Requirements
# Python 3.13+ required
# ============================================================================

# Core AI/ML Libraries
# ============================================================================
openai==2.6.1
anthropic==0.72.0
langchain==1.0.3
langchain-anthropic==1.0.1
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.38

# Vector Database & Embeddings
# ============================================================================
weaviate-client==4.17.0
sentence-transformers>=2.0.0
torch>=2.0.0

# Scientific Computing & ML
# ============================================================================
numpy==2.3.4
scipy==1.16.3
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2

# Deep Learning Support
# ============================================================================
numba==0.62.1
llvmlite==0.45.1

# Data Visualization
# ============================================================================
plotly==6.4.0
umap-learn==0.5.9.post2
pynndescent==0.5.13

# Web UI & Interface
# ============================================================================
gradio>=4.0.0
ipywidgets==8.1.8
widgetsnbextension==4.0.15

# Jupyter Notebook Support
# ============================================================================
jupyter==1.1.1
jupyterlab==4.4.10
jupyter-console==6.6.3
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
notebook==7.4.7
notebook_shim==0.2.4
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1

# HTTP & API Clients
# ============================================================================
requests==2.32.5
requests-toolbelt==1.0.0
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.3
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0

# Web Scraping & Parsing
# ============================================================================
beautifulsoup4==4.14.2
soupsieve==2.8

# Configuration & Environment
# ============================================================================
python-dotenv==1.2.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4

# Data Validation & Serialization
# ============================================================================
dataclasses-json==0.6.7
marshmallow==3.26.1
orjson==3.11.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
validators==0.35.0

# Text Processing & NLP
# ============================================================================
tiktoken==0.12.0
regex==2025.10.23
pytokens==0.2.0

# Database & Storage
# ============================================================================
SQLAlchemy==2.0.44
xxhash==3.6.0
zstandard==0.25.0

# Progress & Logging
# ============================================================================
tqdm==4.67.1
python-json-logger==4.0.0

# Security & Authentication
# ============================================================================
cryptography==46.0.3
Authlib==1.6.5

# Utility Libraries
# ============================================================================
click==8.3.0
arrow==1.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
six==1.17.0
sniffio==1.3.1
anyio==4.11.0
tenacity==9.1.2

# Markup & Templates
# ============================================================================
Jinja2==3.1.6
MarkupSafe==3.0.3
mistune==3.1.4
PyYAML==6.0.3
bleach==6.3.0
defusedxml==0.7.1
tinycss2==1.4.0
webencodings==0.5.1

# Protocol Buffers & Serialization
# ============================================================================
protobuf==6.33.0
ormsgpack==1.11.0
grpcio==1.76.0

# System & Process Management
# ============================================================================
psutil==7.1.2
websocket-client==1.9.0

# Testing & Code Quality
# ============================================================================
pytest==8.4.2
pytest-cov==7.0.0
coverage==7.11.0
black==25.9.0
flake8==7.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0

# Development Tools
# ============================================================================
ipython==9.6.0
jedi==0.19.2
parso==0.8.5
decorator==5.2.1
executing==2.2.1
pure_eval==0.2.3
stack-data==0.6.3
traitlets==5.14.3
backcall>=0.2.0
pickleshare>=0.7.5
prompt_toolkit==3.0.52
wcwidth==0.2.14
Pygments==2.19.2

# Additional Dependencies
# ============================================================================
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
propcache==0.4.1
greenlet==3.2.4
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
annotated-types==0.7.0
docstring_parser==0.17.0
narwhals==2.10.2
lark==1.3.1

# Jupyter Dependencies
# ============================================================================
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
pandocfilters==1.5.1
async-lru==2.0.5
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
babel==2.17.0
fastjsonschema==2.21.2
json5==0.12.1
prometheus_client==0.23.1
Send2Trash==1.8.3
terminado==0.18.1
tornado==6.5.2

# Additional Format Validators
# ============================================================================
fqdn==1.5.1
isoduration==20.11.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
referencing==0.37.0
rpds-py==0.28.0
uri-template==1.3.0
webcolors==25.10.0

# Terminal & Shell
# ============================================================================
pexpect==4.9.0
ptyprocess==0.7.0
appnope==0.1.4

# Platform Tools
# ============================================================================
platformdirs==4.5.0
packaging==25.0
setuptools==80.9.0
distro==1.9.0
deprecation==2.1.0

# Security & Crypto
# ============================================================================
cffi==2.0.0
pycparser==2.23

# System Monitoring
# ============================================================================
comm==0.2.3
nest-asyncio==1.6.0
debugpy==1.8.17
pyzmq==27.1.0
asttokens==3.0.0
matplotlib-inline==0.2.1

# HTTP/Networking
# ============================================================================
h11==0.16.0
iniconfig==2.3.0
jiter==0.11.1
pathspec==0.12.1
pluggy==1.6.0
threadpoolctl==3.6.0


